+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Refactor app structure; delete firebase-releated stuff
 * Request only streams for which user has update permission ([CE-831](https://jira.rfcx.org/browse/CE-831))
 * Add pull_request Gihub action which checks lint errors
+* Add integration tests for uploads routes
 
 <a name="1.0.5"></a>
 ## 1.0.5 (2021-05-26)